move _register_component to utils
